fix: address PRRT_kwDOSJAM6s597ifQ - centralize executor test path helpers
