+++ tests/unit/control/test_executor_validation_fix_cycle.py
@@ -40,6 +40,8 @@ from awf.node.compose_manager import ComposeManager
 from awf.runtime.pr_creator import PullRequestCreator
 from awf.runtime.validation import ValidationRunner
 
+from .executor_paths import _test_worktree_path, _test_worktrees_root
+
 _TEMPLATE = Path(__file__).resolve().parents[3] / "docker" / "compose" / "workspace.base.yml.j2"
 
 
@@ -113,17 +115,6 @@ async def _seed_ready_workspace(
         if create_worktree:
             (_test_worktrees_root(factory) / ws.id).mkdir(parents=True, exist_ok=True)
         return ws.id
-
-
-def _test_worktrees_root(factory: async_sessionmaker[AsyncSession]) -> Path:
-    bind = factory.kw["bind"]
-    database_path = Path(str(bind.url.database))
-    return database_path.parent / "work" / "worktrees"
-
-
-def _test_worktree_path(factory: async_sessionmaker[AsyncSession], workspace_id: str) -> Path:
-    return _test_worktrees_root(factory) / workspace_id
-
 
 def _queue_initial_pass(fake: FakeCommandRunner) -> None:
     """Queue the subprocess results for the initial agent-run + commit
